# apps/hyper-agent/test/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/drives/{key}/file": {
      "get": {
        "operationId": "driveReadFile",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Hex64 drive key",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "drives"
        ]
      },
      "put": {
        "operationId": "driveWriteFile",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Hex64 drive key",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true,
                    "byteLength": 0
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "drives"
        ]
      },
      "delete": {
        "operationId": "driveDeleteEntry",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Hex64 drive key",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "drives"
        ]
      }
    },
...
    "/v1/files/{driveKey}/{path}": {
      "get": {
        "operationId": "readFileRange",
        "parameters": [
          {
            "name": "driveKey",
            "required": true,
            "in": "path",
            "description": "Hex64 drive key",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "Accept-Ranges": {
                "schema": {
                  "type": "string",
                  "example": "bytes"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "public, max-age=31536000, immutable"
                }
              },
              "Content-Range": {
                "schema": {
                  "type": "string",
                  "example": "bytes 0-499/1000"
                }
              }
            },
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "CineReel Hyper Sidecar",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}